fix more int rounding errors and update example

+++ point_controller.cpp
@@ -28,39 +28,39 @@ void point_2d::set_y(int source_y)
 
 // TODO: implement line_2d functions
 
-point_3d::point_3d(int source_x, int source_y, int source_z)
+point_3d::point_3d(float source_x, float source_y, float source_z)
 {
     x = source_x;
     y = source_y;
     z = source_z;
 }
 
-int point_3d::get_x()
+float point_3d::get_x()
 {
     return x;
 }
 
-int point_3d::get_y()
+float point_3d::get_y()
 {
     return y;
 }
 
-int point_3d::get_z()
+float point_3d::get_z()
 {
     return z;
 }
 
-void point_3d::set_x(int source_x)
+void point_3d::set_x(float source_x)
 {
     x = source_x;
 }
 
-void point_3d::set_y(int source_y)
+void point_3d::set_y(float source_y)
 {
     y = source_y;
 }
 
-void point_3d::set_z(int source_z)
+void point_3d::set_z(float source_z)
 {
     z = source_z;
 }
